veracode

+++ centreon/www/include/configuration/configObject/hostgroup/listHostGroup.php
@@ -92,7 +92,7 @@ $rq = "SELECT SQL_CALC_FOUND_ROWS hg_id, hg_name, hg_alias, hg_activate, hg_icon
     "FROM hostgroup " .
     "WHERE {$searchFilterQuery} hg_id NOT IN (SELECT hg_child_id FROM hostgroup_hg_relation) " .
     $acl->queryBuilder('AND', 'hg_id', $hgString) .
-    " ORDER BY hg_name LIMIT " . $num * $limit . ", " . $limit;
+    " ORDER BY hg_name LIMIT " . (int) ($num * $limit) . ", " . (int) $limit;
 $dbResult = $pearDB->prepare($rq);
 $dbResult->execute($mainQueryParameters);
